Nueva pantalla y funcionalidad

+++ FrontAulaRuix/src/app/alumnos/alumnos.component.ts
@@ -3,6 +3,7 @@ import { Usuario } from '../models/usuario';
 import { UsuarioService } from '../service/usuario.service';
 import { ToastrService } from 'ngx-toastr';
 import { TokenService } from '../service/token.service';
+import swal from'sweetalert2';
 
 @Component({
   selector: 'app-alumnos',
@@ -15,6 +16,7 @@ usuarios: Usuario[];
 roles: string[];
 isAdmin = false;
 nombreUsuario: string;
+isProf = false;
 
   constructor(private usuarioService: UsuarioService,
     private toastr: ToastrService,
@@ -28,6 +30,9 @@ nombreUsuario: string;
   this.roles.forEach(rol => {
     if (rol === 'ROLE_ADMIN') {
       this.isAdmin = true;
+    }
+    if (rol === 'ROLE_PROF') {
+      this.isProf = true;
     }
   });
   }
@@ -45,19 +50,28 @@ cargarUsuarios(): void {
 }
 
 borrar(id: number) {
-  this.usuarioService.delete(id).subscribe(
-    data => {
-      this.toastr.success('Usuario Eliminado', 'OK', {
-        timeOut: 3000, positionClass: 'toast-top-center'
-      });
-      this.cargarUsuarios();
-    },
-    err => {
-      this.toastr.error(err.error.mensaje, 'Fail', {
-        timeOut: 3000, positionClass: 'toast-top-center',
-      });
+    swal.fire({
+    title: 'Estas seguro?',
+    text: "¡Una vez borrado no hay vuelta atras!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Sí, borrar'
+  }).then((result) => {
+    if (result.value) {
+      this.usuarioService.delete(id).subscribe(
+        response => {
+          this.cargarUsuarios();
+          swal.fire(
+            'Borrado!',
+            'El usuario a sido borrado con éxito',
+            'success'
+          )
+        }
+      )
     }
-  );
+  })
 }
 
 }
